requirements.txt
Django==1.11.29
django-filter==1.1.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.1.4
drf-nested-routers==0.91